Fix IConvBase ctor
